requirements.txt
Django==4.2
mysqlclient==2.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers